+++ src/View/Helper/LibraryHelper.php
@@ -62,47 +62,6 @@ class LibraryHelper extends Helper
     }
 
     /**
-     * Internal function to generate datepicker and timepicker.
-     *
-     * Bootstrap Datepicker and Moment.js should be installed via Composer.
-     * @param string $input Target field
-     * @param array $options Options for the datepicker
-     * @return string jQuery code
-     * @see http://eonasdan.github.io/bootstrap-datetimepicker Bootstrap 3 Datepicker v4 documentation
-     */
-    protected function buildDatetimepicker(string $input, array $options = []): string
-    {
-        $this->Asset->script([
-            '/vendor/moment/moment-with-locales.min',
-            'MeTools.bootstrap-datetimepicker.min',
-        ], ['block' => 'script_bottom']);
-
-        $this->Asset->css(
-            '/vendor/bootstrap-datetimepicker/css/bootstrap-datetimepicker.min',
-            ['block' => 'css_bottom']
-        );
-
-        $options = optionsParser($options, [
-            'icons' => [
-                'time' => 'fas fa-clock',
-                'date' => 'fas fa-calendar',
-                'up' => 'fas fa-chevron-up',
-                'down' => 'fas fa-chevron-down',
-                'previous' => 'fas fa-chevron-left',
-                'next' => 'fas fa-chevron-right',
-                'today' => 'fas fa-dot-circle',
-                'clear' => 'fas fa-trash',
-                'close' => 'fas fa-times',
-            ],
-            'locale' => substr(I18n::getLocale(), 0, 2) ?: 'en-gb',
-            'showTodayButton' => true,
-            'showClear' => true,
-        ]);
-
-        return sprintf('$("%s").datetimepicker(%s);', $input, json_encode($options->toArray(), JSON_PRETTY_PRINT));
-    }
-
-    /**
      * Before layout callback. beforeLayout is called before the layout is
      *  rendered
      * @return void
@@ -173,46 +132,6 @@ class LibraryHelper extends Helper
     }
 
     /**
-     * Adds a datepicker to the `$input` field.
-     *
-     * To create an input field compatible with datepicker, you should use the
-     *  `datepicker()` method provided by the `FormHelper`.
-     *
-     * Bootstrap Datepicker and Moment.js should be installed via Composer.
-     * @param string $input Target field. Default is `.datepicker`
-     * @param array $options Options for the datepicker
-     * @return void
-     * @deprecated 2.21.1 Use instead the normal functions provided by the most
-     *  modern browsers
-     */
-    public function datepicker(string $input = '', array $options = []): void
-    {
-        deprecationWarning('Deprecated. Use instead the normal functions provided by the most modern browsers');
-
-        $options = optionsParser($options, ['format' => 'YYYY/MM/DD']);
-
-        $this->output[] = $this->buildDatetimepicker($input ?: '.datepicker', $options->toArray());
-    }
-
-    /**
-     * Adds a datetimepicker to the `$input` field.
-     *
-     * To create an input field compatible with datetimepicker, you should use the `datetimepicker()` method provided
-     *  by the `FormHelper`.
-     * Bootstrap Datepicker and Moment.js should be installed via Composer.
-     * @param string $input Target field. Default is `.datetimepicker`
-     * @param array $options Options for the datetimepicker
-     * @return void
-     * @deprecated 2.21.1 Use instead the normal functions provided by the most modern browsers
-     */
-    public function datetimepicker(string $input = '', array $options = []): void
-    {
-        deprecationWarning('Deprecated. Use instead the normal functions provided by the most modern browsers');
-
-        $this->output[] = $this->buildDatetimepicker($input ?: '.datetimepicker', $options);
-    }
-
-    /**
      * Loads all fancybox files
      * @return void
      * @link https://fancyapps.com/fancybox/3 fancybox documentation
@@ -263,26 +182,5 @@ class LibraryHelper extends Helper
         $this->Asset->script('MeTools.slugify', ['block' => 'script_bottom']);
 
         $this->output[] = sprintf('$().slugify("%s", "%s");', $sourceField, $targetField);
-    }
-
-    /**
-     * Adds a timepicker to the `$input` field.
-     *
-     * To create an input field compatible with datepicker, you should use the
-     *  `timepicker()` method provided by the `FormHelper`.
-     *
-     * Bootstrap Datepicker and Moment.js should be installed via Composer.
-     * @param string $input Target field. Default is `.timepicker`
-     * @param array $options Options for the timepicker
-     * @return void
-     * @deprecated 2.21.1 Use instead the normal functions provided by the most modern browsers
-     */
-    public function timepicker(string $input = '', array $options = []): void
-    {
-        deprecationWarning('Deprecated. Use instead the normal functions provided by the most modern browsers');
-
-        $options = optionsParser($options, ['pickTime' => false]);
-
-        $this->output[] = $this->buildDatetimepicker($input ?: '.timepicker', $options->toArray());
     }
 }